+++ backend/internal/models/note.go
@@ -3,10 +3,10 @@ package models
 import "time"
 
 type Note struct {
-	ID        uint      `json:"id" gorm:"primaryKey"`
+	ID        int64     `json:"id" gorm:"primaryKey"`
 	Title     string    `json:"title" gorm:"size:255;not null"`
 	Content   string    `json:"content" gorm:"type:text"`
-	UserID    uint      `json:"user_id" gorm:"not null;index"`
+	UserID    int64     `json:"user_id" gorm:"not null;index"`
 	User      User      `json:"user,omitempty" gorm:"foreignKey:UserID"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
